fix!: `encode::loose_header()` now supports large objects even on 32 bit systems.
Previously, larger than 4GB files wouldn't be supported, which causes problems when
genrating hashes even when streaming data.

+++ gix-object/src/tag/write.rs
@@ -44,8 +44,8 @@ impl crate::WriteTo for Tag {
         Kind::Tag
     }
 
-    fn size(&self) -> usize {
-        b"object".len() + 1 /* space */ + self.target.kind().len_in_hex() + 1 /* nl */
+    fn size(&self) -> u64 {
+        (b"object".len() + 1 /* space */ + self.target.kind().len_in_hex() + 1 /* nl */
             + b"type".len() + 1 /* space */ + self.target_kind.as_bytes().len() + 1 /* nl */
             + b"tag".len() + 1 /* space */ + self.name.len() + 1 /* nl */
             + self
@@ -53,7 +53,7 @@ impl crate::WriteTo for Tag {
             .as_ref()
             .map_or(0, |t| b"tagger".len() + 1 /* space */ + t.size() + 1 /* nl */)
             + 1 /* nl */ + self.message.len()
-            + self.pgp_signature.as_ref().map_or(0, |m| 1 /* nl */ + m.len())
+            + self.pgp_signature.as_ref().map_or(0, |m| 1 /* nl */ + m.len())) as u64
     }
 }
 
@@ -81,8 +81,8 @@ impl<'a> crate::WriteTo for TagRef<'a> {
         Kind::Tag
     }
 
-    fn size(&self) -> usize {
-        b"object".len() + 1 /* space */ + self.target().kind().len_in_hex() + 1 /* nl */
+    fn size(&self) -> u64 {
+        (b"object".len() + 1 /* space */ + self.target().kind().len_in_hex() + 1 /* nl */
             + b"type".len() + 1 /* space */ + self.target_kind.as_bytes().len() + 1 /* nl */
             + b"tag".len() + 1 /* space */ + self.name.len() + 1 /* nl */
             + self
@@ -90,7 +90,7 @@ impl<'a> crate::WriteTo for TagRef<'a> {
                 .as_ref()
                 .map_or(0, |t| b"tagger".len() + 1 /* space */ + t.size() + 1 /* nl */)
             + 1 /* nl */ + self.message.len()
-            + self.pgp_signature.as_ref().map_or(0, |m| 1 /* nl */ + m.len())
+            + self.pgp_signature.as_ref().map_or(0, |m| 1 /* nl */ + m.len())) as u64
     }
 }
